"fix" yaml data structure errors
This is a hacky solution, I intend to do a second pass before calling it done.

Often when using functions like `merge` we get these errors:

	executing "jobTemplate" at <keys>: wrong type for value; expected map[string]interface {}; got map[interface {}]interface {}

This behavior only exhibits itself when using `yaml` variables and does not
exhibit itself when `json` variables are passed.

This is not a limitation for relatively flat varaible structures, but limits
the depth of data that can be operated on in Levant variable files.

This resolves the issue by recursively converting yaml-parsed variable
structures into more concrete types.

Like I said, this is not a great solution, I'm sure it's missing some edge
cases, but it blocks my immediate problem so I'm committing it.